move build_providers_from_urls to common.rs
