+++ crates/e2e-tests/tests/common.rs
@@ -449,3 +449,25 @@ pub async fn send_ckd_request(
     );
     Ok(())
 }
+
+pub fn build_providers_from_urls(
+    urls: &[String],
+    chain_name: &str,
+) -> near_mpc_bounded_collections::NonEmptyBTreeMap<
+    mpc_node_config::foreign_chains::RpcProviderName,
+    mpc_node_config::ForeignChainProviderConfig,
+> {
+    let map: std::collections::BTreeMap<_, _> = urls
+        .iter()
+        .enumerate()
+        .map(|(i, url)| {
+            let cfg = mpc_node_config::ForeignChainProviderConfig {
+                rpc_url: url.clone(),
+                auth: Default::default(),
+            };
+            (format!("mock-{chain_name}-{i}").into(), cfg)
+        })
+        .collect();
+    map.try_into()
+        .unwrap_or_else(|_| panic!("at least one {chain_name} provider must be configured"))
+}
